+++ alt/fint32.go
@@ -58,5 +58,5 @@ func ivalInt32NotEmptyAsString(fi *finfo, rv reflect.Value, addr uintptr) (any, 
 	if v == 0 {
 		return nil, nilValue, true
 	}
-	return strconv.FormatInt(int64(v), 10), nilValue, false
+	return strconv.FormatInt(v, 10), nilValue, false
 }
